moved the create directory to app and to typescript
