+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "dist/app.ts",
   "scripts": {
-    "dev": "node ./src/config/createStorageDirectory.js && nodemon --exec ts-node src/app.ts",
+    "dev": "nodemon --exec ts-node src/app.ts",
     "test": "tsc && jest --detectOpenHandles --forceExit",
     "test:coverage": "tsc && jest --detectOpenHandles --forceExit --coverage",
     "testAuth": "jest --detectOpenHandles --forceExit auth.test.ts",
