+++ src/config/createStorageDirectory.ts
@@ -1,0 +1,13 @@
+const fs = require('fs');
+const path = require('path');
+
+export const createStorageDirectory = () => {
+  const dirPath = path.join(__dirname, '../../storage');
+
+  if (!fs.existsSync(dirPath)) {
+    fs.mkdirSync(dirPath, { recursive: true });
+    console.log('Directory created:', dirPath);
+  } else {
+    console.log('Directory already exists:', dirPath);
+  }
+};
